+++ Schedule.Test/ScheduleEntryTest.cs
