+++ Schedule.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29709.97
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Schedule", "src\Schedule.csproj", "{FCBEE139-FC37-45C3-8E21-F105F832B6DC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Schedule", "Schedule\Schedule.csproj", "{FCBEE139-FC37-45C3-8E21-F105F832B6DC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Schedule.Test", "Schedule.Test\Schedule.Test.csproj", "{0335F77E-4F4F-4118-A007-93C598CB777B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{FCBEE139-FC37-45C3-8E21-F105F832B6DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FCBEE139-FC37-45C3-8E21-F105F832B6DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FCBEE139-FC37-45C3-8E21-F105F832B6DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0335F77E-4F4F-4118-A007-93C598CB777B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0335F77E-4F4F-4118-A007-93C598CB777B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0335F77E-4F4F-4118-A007-93C598CB777B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0335F77E-4F4F-4118-A007-93C598CB777B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
